+++ Assets/Core/Runtime/Localization/LocalizingComponentForText.cs
@@ -8,7 +8,7 @@ namespace KKSFramework.Localization
     /// 변환될 텍스트를 가지고 있는 컴포넌트.
     /// </summary>
     [RequireComponent (typeof (Text))]
-    public class LocalizingComponent : LocalizingComponentBase<Text>
+    public class LocalizingComponentForText : LocalizingComponentBase<Text>
     {
         #region Fields & Property